+++ client/src/components/Smoothie/SmoothieCard.css
@@ -13,7 +13,7 @@
   height: 100%;
   overflow: hidden;
   text-align: center;
-  background:rgb(255, 230, 228)
+  background:rgb(245, 239, 215)
 }
 .card > .img-container > img {
   width: 100%;
