+++ client/src/components/Smoothie/SmoothieDisplay.js
@@ -7,6 +7,8 @@ import Wrapper from "../Wrapper";
 
 import RecipeList from './RecipeList';
 import API from '../../utils/API';
+import JumbotronComponent from "../Jumbotron";
+import StickyFooter from 'react-sticky-footer';
 
 
 
@@ -27,13 +29,29 @@ getRecipes = () => {
 
 componentDidMount() {
   this.getRecipes();
-};
+}; 
 
   render() {
     return (
+      <div>
+      <JumbotronComponent />
       <Wrapper>
         <RecipeList getRecipes={this.getRecipes} dbRecipes={this.state.dbRecipes}/>
       </Wrapper>
+      <StickyFooter
+            bottomThreshold={50}
+            normalStyles={{
+                backgroundColor: "rgba(255,195,77,.8)",
+                padding: "2rem"
+            }}
+            stickyStyles={{
+                backgroundColor: "rgba(255,195,77,.8)",
+                padding: "2rem"
+            }}
+        >
+           copyright @ aliendevelopers  2018
+        </StickyFooter>
+      </div>
     );
   }
 }
