+++ client/src/components/contact/index.js
@@ -1,0 +1,1 @@
+export { default } from "./contact";
