+++ package.json
@@ -55,6 +55,7 @@
     "react-dom": "^16.4.2",
     "react-google-maps": "^9.4.5",
     "react-iframe": "^1.2.0",
+    "react-mail-form": "^1.0.1",
     "react-router-dom": "^4.3.1",
     "react-sticky-footer": "^0.1.0-rc3",
     "reactstrap": "^6.3.1",
